+++ src/greyVest/TagsInput.js
@@ -4,6 +4,7 @@ import { observable } from 'mobx'
 import { observer, inject, useLocalStore } from 'mobx-react'
 import Flex from './Flex'
 import DefaultTag from './Tag'
+import { sanitizeTagWords, splitTagOnComma } from './utils'
 
 let isValidInput = (tag, tags) => !_.isEmpty(tag) && !_.includes(tag, tags)
 
@@ -32,27 +33,16 @@ let TagsInput = forwardRef(
   ) => {
     let containerRef = React.useRef()
     let state = useLocalStore(() => ({ currentInput: '' }))
+    let sanitizeTagFn = sanitizeTagWords(wordsMatchPattern, maxWordsPerTag, maxCharsPerTagWord)
 
-    let words = _.words.convert({ fixed: false })
-    // Convert string to words, take the first maxWordsPerTag, truncate them and convert back to string
-    let sanitizeWords = _.flow(
-      string => words(string, wordsMatchPattern),
-      _.take(maxWordsPerTag),
-      _.map(_.truncate({ length: maxCharsPerTagWord, omission: '' })),
-      _.join(' ')
+    addTags = _.flow(
+      _.trim,
+      tags => splitCommas ? splitTagOnComma(tags) : _.castArray(tags),
+      tags => sanitizeTags ? _.map(sanitizeTagFn, tags) : tags,
+      _.difference(_, tags),
+      addTags
     )
 
-    addTags = splitCommas
-      ? _.flow(
-          _.split(','),
-          _.invokeMap('trim'),
-          _.compact,
-          _.uniq,
-          tags => (sanitizeTags ? _.map(sanitizeWords, tags) : tags),
-          _.difference(_, tags),
-          addTags
-        )
-      : _.flow(_.trim, _.castArray, addTags)
     return (
       <div className={'tags-input'} ref={containerRef} style={{ ...style }}>
         <Flex
